Fixed readdir
ls works with files, but does not list . or .. when the -la flag is
given.

Another bug to note is that when a file is first created, it's size
is not being stored accurately. When a second file is created, the
sizes behave as expected (they should all be zero).

Delete is not fully functional.

Mount and unmount do not persist data.